refactor(reflect): move files around for easier to parse diff
Regenerate `reflect.h2` with the following commands.
```
cppfront -p reflect.h2 -o cpp2reflect.h
mv cpp2reflect.h ../include/
```
