+++ include/cpp2reflect.h
@@ -9,38 +9,38 @@
 
 #include "cpp2util.h"
 
-#line 1 "cpp2reflect.h2"
+#line 1 "reflect.h2"
 
-#line 19 "cpp2reflect.h2"
+#line 19 "reflect.h2"
 namespace cpp2 {
 
 namespace meta {
 
-#line 31 "cpp2reflect.h2"
+#line 31 "reflect.h2"
 class compiler_services;
 
-#line 213 "cpp2reflect.h2"
+#line 213 "reflect.h2"
 class declaration_base;
 
-#line 255 "cpp2reflect.h2"
+#line 255 "reflect.h2"
 class declaration;
 
-#line 337 "cpp2reflect.h2"
+#line 337 "reflect.h2"
 class function_declaration;
 
-#line 424 "cpp2reflect.h2"
+#line 424 "reflect.h2"
 class object_declaration;
 
-#line 460 "cpp2reflect.h2"
+#line 460 "reflect.h2"
 class type_declaration;
 
-#line 595 "cpp2reflect.h2"
+#line 595 "reflect.h2"
 class alias_declaration;
 
-#line 934 "cpp2reflect.h2"
+#line 934 "reflect.h2"
 class value_member_info;
 
-#line 1347 "cpp2reflect.h2"
+#line 1347 "reflect.h2"
 }
 
 }
@@ -48,7 +48,7 @@ class value_member_info;
 
 //=== Cpp2 type definitions and function declarations ===========================
 
-#line 1 "cpp2reflect.h2"
+#line 1 "reflect.h2"
 
 //  Copyright (c) Herb Sutter
 //  SPDX-License-Identifier: CC-BY-NC-ND-4.0
@@ -61,17 +61,17 @@ class value_member_info;
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#line 14 "cpp2reflect.h2"
+#line 14 "reflect.h2"
 //===========================================================================
 //  Reflection and meta
 //===========================================================================
 
-#line 19 "cpp2reflect.h2"
+#line 19 "reflect.h2"
 namespace cpp2 {
 
 namespace meta {
 
-#line 24 "cpp2reflect.h2"
+#line 24 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //  Compiler services
@@ -94,20 +94,20 @@ class compiler_services
         cpp2::in<std::any> data_v
     );
 
-#line 50 "cpp2reflect.h2"
+#line 50 "reflect.h2"
     //  Common API
     //
     public: auto set_metafunction_name(cpp2::in<std::string_view> name, cpp2::in<std::vector<std::string>> args) & -> void;
 
-#line 58 "cpp2reflect.h2"
+#line 58 "reflect.h2"
     public: [[nodiscard]] auto get_metafunction_name() const& -> std::string_view;
 
     public: [[nodiscard]] auto get_argument(cpp2::in<int> index) & -> std::string;
 
-#line 68 "cpp2reflect.h2"
+#line 68 "reflect.h2"
     public: [[nodiscard]] auto get_arguments() & -> std::vector<std::string>;
 
-#line 73 "cpp2reflect.h2"
+#line 73 "reflect.h2"
     public: [[nodiscard]] auto arguments_were_used() const& -> bool;
 
     protected: [[nodiscard]] auto parse_statement(
@@ -115,7 +115,7 @@ class compiler_services
         std::string_view source
     ) & -> auto;
 
-#line 129 "cpp2reflect.h2"
+#line 129 "reflect.h2"
     protected: [[nodiscard]] auto position() const& -> auto;
     private: [[nodiscard]] virtual auto vposition() const -> std::any;
 
@@ -128,24 +128,24 @@ class compiler_services
         cpp2::in<std::string_view> msg
     ) const& -> void;
 
-#line 146 "cpp2reflect.h2"
+#line 146 "reflect.h2"
     public: auto error(cpp2::in<std::string_view> msg) const& -> void;
 
-#line 155 "cpp2reflect.h2"
+#line 155 "reflect.h2"
     //  Enable custom contracts on this object, integrated with compiler output
     //  Unlike .requires, a contract violation stops further processing
     //
     public: auto report_violation(auto const& msg) const& -> void;
 
-#line 163 "cpp2reflect.h2"
+#line 163 "reflect.h2"
     public: [[nodiscard]] auto has_handler() const& -> auto;
     public: virtual ~compiler_services() noexcept;
 public: compiler_services(compiler_services const& that);
 
-#line 164 "cpp2reflect.h2"
+#line 164 "reflect.h2"
 };
 
-#line 167 "cpp2reflect.h2"
+#line 167 "reflect.h2"
 /*
 //-----------------------------------------------------------------------
 //
@@ -182,7 +182,7 @@ type_id: @polymorphic_base @copyable type =
 }
 */
 
-#line 204 "cpp2reflect.h2"
+#line 204 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //  Declarations
@@ -195,7 +195,7 @@ type_id: @polymorphic_base @copyable type =
 class declaration_base
 : public compiler_services {
 
-#line 217 "cpp2reflect.h2"
+#line 217 "reflect.h2"
     public: class node_pointer
      {
         private: std::any n; 
@@ -204,13 +204,13 @@ class declaration_base
 
             T const& n_
         );
-#line 221 "cpp2reflect.h2"
+#line 221 "reflect.h2"
         public: template<typename T> auto operator=(
 
             T const& n_
         ) -> node_pointer& ;
 
-#line 231 "cpp2reflect.h2"
+#line 231 "reflect.h2"
         public: [[nodiscard]] auto operator*() const& -> auto&&;
         public: node_pointer(node_pointer const& that);
 
@@ -218,7 +218,7 @@ public: auto operator=(node_pointer const& that) -> node_pointer& ;
 public: node_pointer(node_pointer&& that) noexcept;
 public: auto operator=(node_pointer&& that) noexcept -> node_pointer& ;
 
-#line 232 "cpp2reflect.h2"
+#line 232 "reflect.h2"
     };
 
     protected: node_pointer n; 
@@ -229,31 +229,31 @@ public: auto operator=(node_pointer&& that) noexcept -> node_pointer& ;
         cpp2::in<compiler_services> s
     );
 
-#line 246 "cpp2reflect.h2"
+#line 246 "reflect.h2"
     private: [[nodiscard]] auto vposition() const -> std::any override;
 
     public: [[nodiscard]] auto print() const& -> std::string;
     public: virtual ~declaration_base() noexcept;
 public: declaration_base(declaration_base const& that);
 
-#line 249 "cpp2reflect.h2"
+#line 249 "reflect.h2"
 };
 
-#line 252 "cpp2reflect.h2"
+#line 252 "reflect.h2"
 //-----------------------------------------------------------------------
 //  All declarations
 //
 class declaration
 : public declaration_base {
 
-#line 259 "cpp2reflect.h2"
+#line 259 "reflect.h2"
     public: explicit declaration(
 
         cpp2::in<declaration_base::node_pointer> n_, 
         cpp2::in<compiler_services> s
     );
 
-#line 268 "cpp2reflect.h2"
+#line 268 "reflect.h2"
     public: [[nodiscard]] auto is_public() const& -> bool;
     public: [[nodiscard]] auto is_protected() const& -> bool;
     public: [[nodiscard]] auto is_private() const& -> bool;
@@ -272,7 +272,7 @@ class declaration
 
     public: [[nodiscard]] auto name() const& -> std::string_view;
 
-#line 289 "cpp2reflect.h2"
+#line 289 "reflect.h2"
     public: [[nodiscard]] auto has_initializer() const& -> bool;
 
     public: [[nodiscard]] auto is_global() const& -> bool;
@@ -315,24 +315,24 @@ public: declaration(declaration const& that);
 
                                                     // this precondition should be sufficient ...
 
-#line 331 "cpp2reflect.h2"
+#line 331 "reflect.h2"
 };
 
-#line 334 "cpp2reflect.h2"
+#line 334 "reflect.h2"
 //-----------------------------------------------------------------------
 //  Function declarations
 //
 class function_declaration
 : public declaration {
 
-#line 341 "cpp2reflect.h2"
+#line 341 "reflect.h2"
     public: explicit function_declaration(
 
         cpp2::in<declaration_base::node_pointer> n_, 
         cpp2::in<compiler_services> s
     );
 
-#line 351 "cpp2reflect.h2"
+#line 351 "reflect.h2"
     public: [[nodiscard]] auto index_of_parameter_named(cpp2::in<std::string_view> s) const& -> int;
     public: [[nodiscard]] auto has_parameter_named(cpp2::in<std::string_view> s) const& -> bool;
     public: [[nodiscard]] auto has_in_parameter_named(cpp2::in<std::string_view> s) const& -> bool;
@@ -369,7 +369,7 @@ class function_declaration
 
     public: [[nodiscard]] auto get_parameters() const& -> std::vector<object_declaration>;
 
-#line 395 "cpp2reflect.h2"
+#line 395 "reflect.h2"
     public: [[nodiscard]] auto is_binary_comparison_function() const& -> bool;
 
     public: auto default_to_virtual() & -> void;
@@ -380,103 +380,103 @@ class function_declaration
     public: function_declaration(function_declaration const& that);
 
 
-#line 418 "cpp2reflect.h2"
+#line 418 "reflect.h2"
 };
 
-#line 421 "cpp2reflect.h2"
+#line 421 "reflect.h2"
 //-----------------------------------------------------------------------
 //  Object declarations
 //
 class object_declaration
 : public declaration {
 
-#line 428 "cpp2reflect.h2"
+#line 428 "reflect.h2"
     public: explicit object_declaration(
 
         cpp2::in<declaration_base::node_pointer> n_, 
         cpp2::in<compiler_services> s
     );
 
-#line 438 "cpp2reflect.h2"
+#line 438 "reflect.h2"
     public: [[nodiscard]] auto is_const() const& -> bool;
     public: [[nodiscard]] auto has_wildcard_type() const& -> bool;
 
     public: [[nodiscard]] auto type() const& -> std::string;
 
-#line 448 "cpp2reflect.h2"
+#line 448 "reflect.h2"
     public: [[nodiscard]] auto initializer() const& -> std::string;
     public: object_declaration(object_declaration const& that);
 
 
-#line 454 "cpp2reflect.h2"
+#line 454 "reflect.h2"
 };
 
-#line 457 "cpp2reflect.h2"
+#line 457 "reflect.h2"
 //-----------------------------------------------------------------------
 //  Type declarations
 //
 class type_declaration
 : public declaration {
 
-#line 464 "cpp2reflect.h2"
+#line 464 "reflect.h2"
     public: explicit type_declaration(
 
         cpp2::in<declaration_base::node_pointer> n_, 
         cpp2::in<compiler_services> s
     );
 
-#line 474 "cpp2reflect.h2"
+#line 474 "reflect.h2"
     public: auto reserve_names(cpp2::in<std::string_view> name, auto&& ...etc) const& -> void;
 
-#line 486 "cpp2reflect.h2"
+#line 486 "reflect.h2"
     public: [[nodiscard]] auto is_polymorphic() const& -> bool;
     public: [[nodiscard]] auto is_final() const& -> bool;
     public: [[nodiscard]] auto make_final() & -> bool;
 
     public: [[nodiscard]] auto get_member_functions() const& -> std::vector<function_declaration>;
 
-#line 501 "cpp2reflect.h2"
+#line 501 "reflect.h2"
     public: [[nodiscard]] auto get_member_functions_needing_initializer() const& -> std::vector<function_declaration>;
 
-#line 516 "cpp2reflect.h2"
+#line 516 "reflect.h2"
     public: [[nodiscard]] auto get_member_objects() const& -> std::vector<object_declaration>;
 
-#line 526 "cpp2reflect.h2"
+#line 526 "reflect.h2"
     public: [[nodiscard]] auto get_member_types() const& -> std::vector<type_declaration>;
 
-#line 536 "cpp2reflect.h2"
+#line 536 "reflect.h2"
     public: [[nodiscard]] auto get_member_aliases() const& -> std::vector<alias_declaration>;
 
-#line 546 "cpp2reflect.h2"
+#line 546 "reflect.h2"
     public: [[nodiscard]] auto get_members() const& -> std::vector<declaration>;
 struct query_declared_value_set_functions_ret { bool out_this_in_that; bool out_this_move_that; bool inout_this_in_that; bool inout_this_move_that; };
 
 
 
-#line 556 "cpp2reflect.h2"
+#line 556 "reflect.h2"
     public: [[nodiscard]] auto query_declared_value_set_functions() const& -> query_declared_value_set_functions_ret;
 
-#line 571 "cpp2reflect.h2"
+#line 571 "reflect.h2"
     public: auto add_member(cpp2::in<std::string_view> source) & -> void;
 
-#line 585 "cpp2reflect.h2"
+#line 585 "reflect.h2"
     public: auto remove_marked_members() & -> void;
     public: auto remove_all_members() & -> void;
 
     public: auto disable_member_function_generation() & -> void;
     public: type_declaration(type_declaration const& that);
 
-#line 589 "cpp2reflect.h2"
+#line 589 "reflect.h2"
 };
 
-#line 592 "cpp2reflect.h2"
+#line 592 "reflect.h2"
 //-----------------------------------------------------------------------
 //  Alias declarations
 //
 class alias_declaration
 : public declaration {
 
-#line 599 "cpp2reflect.h2"
+#line 599 "reflect.h2"
     public: explicit alias_declaration(
 
         cpp2::in<declaration_base::node_pointer> n_, 
@@ -485,10 +485,10 @@ class alias_declaration
     public: alias_declaration(alias_declaration const& that);
 
 
-#line 608 "cpp2reflect.h2"
+#line 608 "reflect.h2"
 };
 
-#line 611 "cpp2reflect.h2"
+#line 611 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //  Metafunctions - these are hardwired for now until we get to the
@@ -503,7 +503,7 @@ class alias_declaration
 //
 auto add_virtual_destructor(meta::type_declaration& t) -> void;
 
-#line 629 "cpp2reflect.h2"
+#line 629 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //      "... an abstract base class defines an interface ..."
@@ -518,7 +518,7 @@ auto add_virtual_destructor(meta::type_declaration& t) -> void;
 //
 auto interface(meta::type_declaration& t) -> void;
 
-#line 668 "cpp2reflect.h2"
+#line 668 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "C.35: A base class destructor should be either public and
@@ -540,7 +540,7 @@ auto interface(meta::type_declaration& t) -> void;
 //
 auto polymorphic_base(meta::type_declaration& t) -> void;
 
-#line 712 "cpp2reflect.h2"
+#line 712 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "... A totally ordered type ... requires operator<=> that
@@ -566,7 +566,7 @@ auto ordered_impl(
     cpp2::in<std::string_view> ordering// must be "strong_ordering" etc.
 ) -> void;
 
-#line 756 "cpp2reflect.h2"
+#line 756 "reflect.h2"
 //-----------------------------------------------------------------------
 //  ordered - a totally ordered type
 //
@@ -574,19 +574,19 @@ auto ordered_impl(
 //
 auto ordered(meta::type_declaration& t) -> void;
 
-#line 766 "cpp2reflect.h2"
+#line 766 "reflect.h2"
 //-----------------------------------------------------------------------
 //  weakly_ordered - a weakly ordered type
 //
 auto weakly_ordered(meta::type_declaration& t) -> void;
 
-#line 774 "cpp2reflect.h2"
+#line 774 "reflect.h2"
 //-----------------------------------------------------------------------
 //  partially_ordered - a partially ordered type
 //
 auto partially_ordered(meta::type_declaration& t) -> void;
 
-#line 783 "cpp2reflect.h2"
+#line 783 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "A value is ... a regular type. It must have all public
@@ -605,7 +605,7 @@ auto partially_ordered(meta::type_declaration& t) -> void;
 //
 auto copyable(meta::type_declaration& t) -> void;
 
-#line 820 "cpp2reflect.h2"
+#line 820 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //  basic_value
@@ -615,7 +615,7 @@ auto copyable(meta::type_declaration& t) -> void;
 //
 auto basic_value(meta::type_declaration& t) -> void;
 
-#line 845 "cpp2reflect.h2"
+#line 845 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "A 'value' is a totally ordered basic_value..."
@@ -628,13 +628,13 @@ auto basic_value(meta::type_declaration& t) -> void;
 //
 auto value(meta::type_declaration& t) -> void;
 
-#line 861 "cpp2reflect.h2"
+#line 861 "reflect.h2"
 auto weakly_ordered_value(meta::type_declaration& t) -> void;
 
-#line 867 "cpp2reflect.h2"
+#line 867 "reflect.h2"
 auto partially_ordered_value(meta::type_declaration& t) -> void;
 
-#line 874 "cpp2reflect.h2"
+#line 874 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "By definition, a `struct` is a `class` in which members
@@ -662,7 +662,7 @@ auto partially_ordered_value(meta::type_declaration& t) -> void;
 //
 auto cpp2_struct(meta::type_declaration& t) -> void;
 
-#line 917 "cpp2reflect.h2"
+#line 917 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "C enumerations constitute a curiously half-baked concept. ...
@@ -692,7 +692,7 @@ auto basic_enum(
     cpp2::in<bool> bitwise
     ) -> void;
 
-#line 1123 "cpp2reflect.h2"
+#line 1123 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //    "An enum[...] is a totally ordered value type that stores a
@@ -704,7 +704,7 @@ auto basic_enum(
 //
 auto cpp2_enum(meta::type_declaration& t) -> void;
 
-#line 1149 "cpp2reflect.h2"
+#line 1149 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "flag_enum expresses an enumeration that stores values
@@ -717,7 +717,7 @@ auto cpp2_enum(meta::type_declaration& t) -> void;
 //
 auto flag_enum(meta::type_declaration& t) -> void;
 
-#line 1181 "cpp2reflect.h2"
+#line 1181 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //     "As with void*, programmers should know that unions [...] are
@@ -744,14 +744,14 @@ auto flag_enum(meta::type_declaration& t) -> void;
 
 auto cpp2_union(meta::type_declaration& t) -> void;
 
-#line 1337 "cpp2reflect.h2"
+#line 1337 "reflect.h2"
 //-----------------------------------------------------------------------
 //
 //  print - output a pretty-printed visualization of t
 //
 auto print(cpp2::in<meta::type_declaration> t) -> void;
 
-#line 1347 "cpp2reflect.h2"
+#line 1347 "reflect.h2"
 }
 
 }
