--- source/cpp2reflect.h
@@ -1,1 +1,0 @@
-#include "../include/cpp2reflect.h"
